+++ packages/compiler-cli/src/ngtsc/typecheck/index.ts
@@ -9,7 +9,6 @@
 export {ReusedProgramStrategy} from './src/augmented_program';
 export {FileTypeCheckingData, TemplateTypeCheckerImpl} from './src/checker';
 export {TypeCheckContextImpl} from './src/context';
-export {isTemplateDiagnostic, TemplateDiagnostic} from './src/diagnostics';
 export {TypeCheckProgramHost} from './src/host';
 export {TypeCheckShimGenerator} from './src/shim';
 export {typeCheckFilePath} from './src/type_check_file';
