+++ packages/compiler-cli/src/ngtsc/typecheck/src/dom.ts
@@ -11,8 +11,9 @@ import * as ts from 'typescript';
 
 import {ErrorCode, ngErrorCode} from '../../diagnostics';
 import {TemplateId} from '../api';
+import {makeTemplateDiagnostic, TemplateDiagnostic} from '../diagnostics';
 
-import {makeTemplateDiagnostic, TemplateDiagnostic, TemplateSourceResolver} from './diagnostics';
+import {TemplateSourceResolver} from './diagnostics';
 
 const REGISTRY = new DomElementSchemaRegistry();
 const REMOVE_XHTML_REGEX = /^:xhtml:/;
