+++ packages/compiler-cli/src/ngtsc/typecheck/src/source.ts
@@ -8,7 +8,9 @@
 
 import {AbsoluteSourceSpan, ParseLocation, ParseSourceFile, ParseSourceSpan} from '@angular/compiler';
 import * as ts from 'typescript';
+
 import {TemplateId, TemplateSourceMapping} from '../api';
+import {getTemplateId} from '../diagnostics';
 
 import {TemplateSourceResolver} from './diagnostics';
 import {computeLineStartsMap, getLineAndCharacterFromPosition} from './line_mappings';
@@ -80,29 +82,4 @@ export class TemplateSourceManager implements TemplateSourceResolver {
     const templateSource = this.templateSources.get(id)!;
     return templateSource.toParseSourceSpan(span.start, span.end);
   }
-}
-
-const TEMPLATE_ID = Symbol('ngTemplateId');
-const NEXT_TEMPLATE_ID = Symbol('ngNextTemplateId');
-
-interface HasTemplateId {
-  [TEMPLATE_ID]: TemplateId;
-}
-
-interface HasNextTemplateId {
-  [NEXT_TEMPLATE_ID]: number;
-}
-
-function getTemplateId(node: ts.ClassDeclaration&Partial<HasTemplateId>): TemplateId {
-  if (node[TEMPLATE_ID] === undefined) {
-    node[TEMPLATE_ID] = allocateTemplateId(node.getSourceFile());
-  }
-  return node[TEMPLATE_ID]!;
-}
-
-function allocateTemplateId(sf: ts.SourceFile&Partial<HasNextTemplateId>): TemplateId {
-  if (sf[NEXT_TEMPLATE_ID] === undefined) {
-    sf[NEXT_TEMPLATE_ID] = 1;
-  }
-  return (`tcb${sf[NEXT_TEMPLATE_ID]!++}`) as TemplateId;
 }
